requirements.txt
# Function dependencies, for example:
# package>=version

line-bot-sdk==3.10.2
requests==2.31.0
git+https://github.com/ozgur/python-firebase
firebase_admin